Enhance engine robustness, worksheet I/O, tests
Multiple improvements across engine, UI, core, and tests: add transactional gating, timed write/read, executable injection and SIGPIPE protection to EngineClient; propagate column types in protocol/bridge and preserve project Excel/text semantics; add AtomicFile for safe atomic saves and use it for CSV/XLSX exports; enforce contiguous finite observations for time-series/control-chart ops via OrderedNumericValues and Columns.Rows(requireContiguous); tighten numeric parsing with NumericCell and reject non-finite cells; adjust moving-average to trailing windows and label forward-selection consistently; expose KMeans convergence and fix several numeric/overflow/formatting edge cases. Also add comprehensive audit tests (mac + smoke) and small UI fixes for worksheet editing, import header option, conflict handling and editor commit behavior; update README accordingly.

+++ src/StatStudio.Core/Statistics/HypothesisTests.cs
@@ -123,7 +123,7 @@ public static class HypothesisTests
         if (n1 <= 0 || n2 <= 0 || x1 < 0 || x1 > n1 || x2 < 0 || x2 > n2 || conf is <= 0 or >= 1)
             throw new ArgumentException("Proportion inputs require 0 ≤ x ≤ n, n > 0, and valid confidence.");
         double p1 = (double)x1 / n1, p2 = (double)x2 / n2, diff = p1 - p2;
-        double pPool = (double)(x1 + x2) / (n1 + n2);
+        double pPool = ((double)x1 + x2) / ((double)n1 + n2);
         double seTest = Math.Sqrt(pPool * (1 - pPool) * (1.0 / n1 + 1.0 / n2));
         double z = seTest > 0 ? diff / seTest : double.NaN;
         double p = PFromZ(z, alt);
